requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-chartit==0.1
django-tastypie==0.12.1
ecdsa==0.13
Fabric==1.10.1
gunicorn==19.1.1
lxml==3.4.4
MySQL-python==1.2.5
paramiko==1.15.2
psycopg2==2.5.1
py-mysql2pgsql==0.1.5
pycrypto==2.6.1
PyMySQL==0.6.6
python-dateutil==2.4.2
python-mimeparse==0.1.4
PyYAML==3.11
simplejson==3.7.3
six==1.9.0
static3==0.6.1
termcolor==1.1.0
wheel==0.24.0
whitenoise==1.0.6
